Fix Auto-Rendered Items
This commit comes down to a sigle line of code, line 92 of BakedInfItemModel, which caused the auto-rendered items to not render at all. This was a suprisingly difficult bug to debug, as it manifested in a manner that seemingly makes no sense, where a BakedModel would be created by the rendering pipeline, yet somehow the method getQuads was never called on it. Anyways, I digress.