Fix ItemBlocks
This may not be the best way as to associate ItemBlocks with Items, however it does maintain the tight association between the two. It might have been cleaner to require all ItemBlocks to be registered in the mod Item registry file, but I digress, as this is working for now.