Backport Memory Fix
Also updated the mappings to the latest stable mappings for 1.10.2. The mappings update might not have been the best idea to combine with the backport, but whatever.

This commit contains breaking changes that were deemed unavoidable. Luckily it's only a single method that's been changed that should be easy to update.